refactor: 폴더명 및 파일명 정리
