+++ src/containers/CardListModal/index.jsx
@@ -1,7 +1,7 @@
 import { useCallback, useContext } from 'react';
 import CardCompany from 'components/CardCompany';
 import ToastModal from 'components/ToastModal/ToastModal';
-import { CARD_COMPANIES } from 'lib/constants/cardDomain';
+import { CARD_COMPANIES } from 'lib/constants';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
